Implemented sound, tweaked falling
BGM plays; sound effects for running, jumping, landing; player now
properly enters falling state when walking off a platform or spawning.

Running sound is a little weird rn; it plays to completion even if you
jump or stop moving right after the first step, for example. Not a huge
priority, but I'll see if I can't find a way to fix it.